+++ CloudMicroservices/PeripheryInputChannelMessage.cs
@@ -2,7 +2,7 @@
 {
     public class PeripheryInputChannelMessage : PeripheryChannelMessage
     {
-        public ulong CorrelationId { get; set; }
+        public uint CorrelationId { get; set; }
 
         public override string ToString()
         {
